examples/microservices/orders/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
psycopg2-binary==2.9.*
redis==5.2.*
pydantic==2.*
